Sit back and relax... Ninja draws his sword.
All forum topics and posts are now sent to the adress specified in settings.py. Furthermore, it supports threaded e-mail ;)

Later, quotation will be added to forum and e-mail will be sent just as e-mail clients do. We will be able to watch our greatest via KMail or Thunderbird easily ;)